+++ db/migrate/001_add_accept_terms_at_to_user.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 class AddAcceptTermsAtToUser < ActiveRecord::Migration[4.2]
   def change
     add_column :users, :accept_terms_at, :datetime