+++ test/unit/i18n_test.rb
@@ -1,4 +1,6 @@
-require File.expand_path('../../test_helper', __FILE__)
+# frozen_string_literal: true
+
+require File.expand_path '../test_helper', __dir__
 
 class I18nTest < ActiveSupport::TestCase
   include Redmine::I18n